Change repository structure according to go best practices
