+++ pkg/flaki/transport/http/middleware.go
@@ -1,0 +1,37 @@
+package http
+
+import (
+	"net/http"
+
+	opentracing "github.com/opentracing/opentracing-go"
+	otag "github.com/opentracing/opentracing-go/ext"
+)
+
+// Middleware on http transport.
+type Middleware func(http.Handler) http.Handler
+
+// MakeTracingMiddleware try to extract an existing span from the HTTP headers. It it exists, we
+// continue the span, if not we create a new one.
+func MakeTracingMiddleware(tracer opentracing.Tracer, operationName string) Middleware {
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+
+			var sc, err = tracer.Extract(opentracing.HTTPHeaders, opentracing.HTTPHeadersCarrier(r.Header))
+
+			var span opentracing.Span
+			if err != nil {
+				span = tracer.StartSpan(operationName)
+			} else {
+				span = tracer.StartSpan(operationName, opentracing.ChildOf(sc))
+			}
+			defer span.Finish()
+
+			// Set tags.
+			otag.Component.Set(span, "flaki-service")
+			span.SetTag("transport", "http")
+			otag.SpanKindRPCServer.Set(span)
+
+			next.ServeHTTP(w, r.WithContext(opentracing.ContextWithSpan(r.Context(), span)))
+		})
+	}
+}
