+++ script/build.sh
@@ -36,19 +36,13 @@ if [ -z ${ENV} ]; then
 fi
 
 # Directories.
-FLATBUF_DIR="./service/transport/flatbuffer"
+FLATBUF_DIR="./pkg/flaki/transport/flatbuffer"
 
 # Delete the old dirs.
 echo "==> Removing old directories..."
 rm -f bin/*
 mkdir -p bin/
 rm -f "$FLATBUF_DIR"/fb/*
-
-# Get the git commit.
-GIT_COMMIT="$(git rev-parse HEAD)"
-
-# Override the variables GitCommit and Environment in the main package.
-LD_FLAGS="-X main.GitCommit=${GIT_COMMIT} -X main.Environment=${ENV}"
 
 # Flatbuffers.
 echo
@@ -60,10 +54,18 @@ ls -hl "$FLATBUF_DIR"/fb
 echo
 echo "==> Build:"
 
+cd cmd
+
+# Get the git commit.
+GIT_COMMIT="$(git rev-parse HEAD)"
+
+# Override the variables GitCommit and Environment in the main package.
+LD_FLAGS="-X main.GitCommit=${GIT_COMMIT} -X main.Environment=${ENV}"
+
 #export CGO_ENABLED="0"
 
-go build -ldflags "$LD_FLAGS" -o bin/flaki_service
+go build -ldflags "$LD_FLAGS" -o ../bin/flaki_service
 echo "Build commit '${GIT_COMMIT}' for '${ENV}' environment."
-ls -hl bin/
+ls -hl ../bin/
 
 exit 0
