--- service/transport/flatbuffer/fb/EmptyRequest.go
@@ -1,34 +1,0 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
-package fb
-
-import (
-	flatbuffers "github.com/google/flatbuffers/go"
-)
-
-type EmptyRequest struct {
-	_tab flatbuffers.Table
-}
-
-func GetRootAsEmptyRequest(buf []byte, offset flatbuffers.UOffsetT) *EmptyRequest {
-	n := flatbuffers.GetUOffsetT(buf[offset:])
-	x := &EmptyRequest{}
-	x.Init(buf, n+offset)
-	return x
-}
-
-func (rcv *EmptyRequest) Init(buf []byte, i flatbuffers.UOffsetT) {
-	rcv._tab.Bytes = buf
-	rcv._tab.Pos = i
-}
-
-func (rcv *EmptyRequest) Table() flatbuffers.Table {
-	return rcv._tab
-}
-
-func EmptyRequestStart(builder *flatbuffers.Builder) {
-	builder.StartObject(0)
-}
-func EmptyRequestEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {
-	return builder.EndObject()
-}
